Voronoi world map layout with MapGenerator, Polygon boundaries, and Road rename
- Add MapGenerator using Voronoi tessellation to place cities, assign country
  boundaries via directed boundary walk, and generate MST roads
- Replace Rectangle boundaries on City/Country with Polygon (IReadOnlyCollection<Point>)
- Rename TravelRoute to Road; fold road generation into GenerateGeographyCommand
- GenerateGeographyCommand now calls MapGenerator.Generate for layout; LLM only
  names world, countries, cities, and roads - distances computed from city centers
- Add WorldWidth/WorldHeight/CityTileSize to WorldGenerationDefaults
- Remove GenerateTravelRoutesCommandHandler; gut GenerateTravelRoutesCommand.cs
- Add migration for Road table rename and Polygon boundary schema change

+++ TRPG/Models/Point.cs
@@ -1,3 +1,3 @@
 ﻿namespace TRPG.Models;
 
-internal record Point(int X, int Y);
+internal record Point(double X, double Y);
